+++ addons/account/static/src/xml/account_payment.xml
@@ -2,7 +2,7 @@
 
 <templates xml:space="preserve">
 
-    <t t-name="ShowPaymentInfo">
+    <t t-name="account.AccountPaymentField" owl="1">
         <div>
             <t t-if="outstanding">
                 <div>
@@ -10,7 +10,7 @@
                 </div>
             </t>
             <table style="width:100%;">
-                <t t-foreach="lines" t-as="line">
+                <t t-foreach="lines" t-as="line" t-key="line_index">
                     <tr>
                     <t t-if="outstanding">
                         <td>
@@ -20,7 +20,8 @@
                                t-att-data-id="line.id"
                                style="margin-right: 0px; padding-left: 5px; padding-right: 5px;"
                                href="#"
-                               data-bs-toggle="tooltip">Add</a>
+                               data-bs-toggle="tooltip"
+                               t-on-click.prevent="() => this.assignOutstandingCredit(line.id)">Add</a>
                         </td>
                         <td style="max-width: 11em;">
                             <a t-att-title="line.date"
@@ -35,7 +36,7 @@
                     </t>
                     <t t-if="!outstanding">
                         <td>
-                           <a role="button" tabindex="0" class="js_payment_info fa fa-info-circle" t-att-index="line.index" style="margin-right:5px;" aria-label="Info" title="Journal Entry Info" data-bs-toggle="tooltip"></a>
+                           <a role="button" tabindex="0" class="js_payment_info fa fa-info-circle" t-att-index="line_index" style="margin-right:5px;" aria-label="Info" title="Journal Entry Info" data-bs-toggle="tooltip" t-on-click.stop="(ev) => this.onInfoClick(ev, line_index)"></a>
                         </td>
                         <td t-if="!line.is_exchange">
                             <i class="o_field_widget text-start o_payment_label">Paid on <t t-esc="line.date"></t></i>
@@ -43,27 +44,15 @@
                         <td t-if="line.is_exchange" colspan="2">
                             <i class="o_field_widget text-start text-muted text-start">
                                 <span class="oe_form_field oe_form_field_float oe_form_field_monetary fw-bold">
-                                    <t t-if="line.position === 'before'">
-                                        <t t-esc="line.currency"/>
-                                    </t>
-                                    <t t-esc="line.amount"/>
-                                    <t t-if="line.position === 'after'">
-                                        <t t-esc="line.currency"/>
-                                    </t>
+                                    <t t-esc="line.amount_formatted"/>
                                 </span>
-                                <span>Exchange Difference</span>
+                                <span> Exchange Difference</span>
                             </i>
                         </td>
                     </t>
                     <td t-if="!line.is_exchange" style="text-align:right;">
                         <span class="oe_form_field oe_form_field_float oe_form_field_monetary" style="margin-left: -10px;">
-                            <t t-if="line.position === 'before'">
-                                <t t-esc="line.currency"/>
-                            </t>
-                            <t t-esc="line.amount"/>
-                            <t t-if="line.position === 'after'">
-                                <t t-esc="line.currency"/>
-                            </t>
+                            <t t-esc="line.amount_formatted"/>
                         </span>
                     </td>
                     </tr>
@@ -72,38 +61,43 @@
         </div>
     </t>
 
-    <t t-name="PaymentPopOver">
-        <div>
-            <table>
-                <tr>
-                    <td><strong>Amount: </strong></td>
-                    <td>
-                        <t t-esc="amount_company_currency"></t>
-                        <t t-if="amount_foreign_currency">
-                            (<span class="fa fa-money"/> <t t-esc="amount_foreign_currency"/>)
-                        </t>
-                    </td>
-                </tr>
-                 <tr>
-                    <td><strong>Memo: </strong></td>
-                    <td>
-                        <div style="width: 200px; word-wrap: break-word">
-                            <t t-esc="ref"/>
-                        </div>
-                    </td>
-                </tr>
-                <tr>
-                    <td><strong>Date: </strong></td>
-                    <td><t t-esc="date"/></td>
-                </tr>
-                <tr>
-                    <td><strong>Journal: </strong></td>
-                    <td><t t-esc="journal_name"/><span t-if="payment_method_name"> (<t t-esc="payment_method_name"/>)</span></td>
-                </tr>
-            </table>
+    <t t-name="account.AccountPaymentPopOver" owl="1">
+        <div class="account_payment_popover">
+            <h3 t-if="props.title" class="o_popover_header"><t t-esc="props.title"/></h3>
+            <div class="px-2">
+                <div>
+                    <table>
+                        <tr>
+                            <td><strong>Amount: </strong></td>
+                            <td>
+                                <t t-esc="props.amount_company_currency"></t>
+                                <t t-if="props.amount_foreign_currency">
+                                    (<span class="fa fa-money"/> <t t-esc="props.amount_foreign_currency"/>)
+                                </t>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td><strong>Memo: </strong></td>
+                            <td>
+                                <div style="width: 200px; word-wrap: break-word">
+                                    <t t-esc="props.ref"/>
+                                </div>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td><strong>Date: </strong></td>
+                            <td><t t-esc="props.date"/></td>
+                        </tr>
+                        <tr>
+                            <td><strong>Journal: </strong></td>
+                            <td><t t-esc="props.journal_name"/><span t-if="props.payment_method_name"> (<t t-esc="props.payment_method_name"/>)</span></td>
+                        </tr>
+                    </table>
+                </div>
+                <button class="btn btn-sm btn-primary js_unreconcile_payment float-start" t-if="!props.is_exchange" style="margin-top:5px; margin-bottom:5px;" groups="account.group_account_invoice" t-on-click="() => props._onRemoveMoveReconcile(props.move_id, props.partial_id)">Unreconcile</button>
+                <button class="btn btn-sm btn-secondary js_open_payment float-end" style="margin-top:5px; margin-bottom:5px;" t-on-click="() => props._onOpenMove(props.move_id)">View</button>
+            </div>
         </div>
-        <button class="btn btn-sm btn-primary js_unreconcile_payment float-start" t-if="!is_exchange" t-att-partial-id="partial_id" t-att-payment-id="payment_id" t-att-move-id="move_id" style="margin-top:5px; margin-bottom:5px;" groups="account.group_account_invoice">Unreconcile</button>
-        <button class="btn btn-sm btn-secondary js_open_payment float-end" t-att-move-id="move_id" style="margin-top:5px; margin-bottom:5px;">View</button>
     </t>
 
 </templates>
